[carnelian] Implement update area tracking
Create an iterator to iterate over all included pixels in an update
area. The use case for this iteration is to allow canvas to update only
the portions of the pixel map that have changed.

One specific use case for this feature is to avoid redrawing the entire
background of the recovery UI when just the two strings have changed.
The single call to fill_rect() that redraws the background can have its
pixel iteration limited to the area that the labels occupied before
updating.

This update area feature is off by default, as it isn’t generally
useful to programs using canvas.

I feel this change is in the spirit of the testing hackathon, as the
new feature includes a number of tests. If the testability review
disagrees, this can wait another week to be merged.

Change-Id: I8b676a5e1715a6fe3c41a4be470c2d1aadf2cce5