[trace] Allow launching apps from `trace record` in a nested environment.
Test: NestedTestEnvironment.Test
      Spec.{DecodingErrors, DecodeEmpty, DecodeEnvironment}

This helps benchmarks on memory usage, uniquely identify the tested component.

For more context: To get the memory usage of a component we need to iterate
throught the running processes (see TaskEnumerator in task-utils/walker.h),
until we find the one with the expected name (e.g. `ledger.cmx` for ledger).
Using the process handle we can then find its memory usage.

However, if more than one process with that name is running, it is
impossible to determine which is the right one. This CL solves this by
creating a nested environment: now both the benchmark component and the
tested one are under the same environment, named "trace_record_environment".
Using this information, it will be possible to correctly identify the process
that shares the same parent with the benchmarks.

Change-Id: I2f3ef0baf769990d6498f0bd9a406ea052d2b0da