[fuchsia-runtime] add HandleId type for startup handles
In C, PA_HND creates a handle id that contains a handle type an an
optional argument. In rust, the create_handle_id function did the same
thing as PA_HND, but it returned a simple u32, which couldn't be used
with any of the rust functions that took a variant of HandleType. The
equivalent C functions that take a handle type also take that optional
argument.

This creates a type HandleId, which is a struct with a u32 element,
created by HandleId::create (renamed from create_handle_id). Functions
that took a HandleType are changed to take a HandleId instead. A
From<HandleType> implementation is provided for HandleId, which sets the
argument to 0.

Change-Id: I31293e7f6d783317ce2701eb9ad2205c64f82ac7