[modular][testing] Rewrite the session_shell integration test with modular test harness.
See peridot/tests/session_shell/session_shell_test_session_shell.cc for
the test this is trying to replace.

A few things to note here:
- Factoring some of the boilerplate used to intercept the session shell
and initialize the mock into a helper function is a possibly
controversial choice. I think it's worthwhile here, since it's not
really necessary for understanding the test.

- The mock session shell used here is not in its final form, as noted in
the TODO. There is another issue (MF-389) that will refactor it into a
utility class.

- The final test, which tests whether DetachView() is called during
logout, is not implemented yet. It will probably add some significant
complexity / length since it will have to intercept and mock base shell, so I
decided to do that in a future change (this change is already too big).

MF-399

Change-Id: I389d78b6db3c345e5f085826492dc390ce9f9fad