[kernel][exception] fix race in exception-test
This adds some waits for ZX_THREAD_TERMINATED signals; previously there
was a race where the process exit() could kill the thread first which
would cause the unittest crash handler to never see the expected
exception.

This race currently seems very unlikely to show up, but an upcoming
patch makes a change to kill/exception behavior which makes this much
more apparent so we need to fix this first.

ZX-4132

Change-Id: I5e1dce196631730074651449ede881d6aa458b84