[debugger] Differentiate async and known-unknown regs.
Register values can have three states:
  - Known.
  - Known unknown (register is not available).
  - Requires async query.
Previously the first two states were conflated, requiring an async call
to differentiate unavailable and asynchronously availab.e

We are moving to keeping all known general registers on the frame so
all general registers will either be known or known-unknown. This is not
hooked up yet, but when it is, known non-general registers (e.g. vector
and debug) will require async calls to get the value for the top stack
frame.

No user-visible change, should be covered by existing tests. One test
was updated to account for the new synchronously known to be unknown
state.

Change-Id: I5ba58bc100e44fae5fd85a8f93cb5a49574e43b9