[debugger] Add a perf testing target.
Adds a framework and test target for adding zxdb perf tests.

This framework is very simple. It can run on both the host and the
target, unlike the existing //zircon/syste,/ulib/perftest which is quite
tied to the Zircon system. The test primitives are based on those in
Chrome.

Adds an empty module load perf test. This is to validate that the output
has the correct format but currently tests nothing.

Change-Id: Ib1517314d4f43f68d4ffc06575f048d524dc9b88