[zircon][lib][fvm] Utility method for resize.
On the next patch this methods will be used to verify that the
conditions are met for growing the fvm. Essentially introduces
invariants based on the metadata allocated size, in terms of the
maximum addressable partition, slices, and allocatable slices.

Some of these methods takes the current disk size as parameter,
which is useful when growing in multiple steps.

TEST=format-info-test

Change-Id: Id68e77f1b83a7df4b603c2cdb337fb8609b7e1c4