[cobalt-client] Declare operator= along with copy constructor
It's deprecated practice in C++ to declare the copy constructor but then
use an implicitly defaulted operator=.  If the copy constructor is
explicit, then operator= must be explicit.  In HistogramOptions the copy
constructor is declared so it can be defined = default out of line.  So
do the same for operator=.  (It could as well be declared = default or
both could be, or both could be undeclared and implicitly defaulted.
Note the default constructor is explicitly declared = default,
i.e. inlined.  But avoiding excessive inlining is sensible for both.)

Change-Id: I222c080ea530d52359caf8437f9c0e4f9f564326