[e2e] Rename src/end_to_end_tests to src/tests.
What makes a test end to end is that it is configured to be run
as host side test and that it executes the end to end testing
machinery. It's not confined to be at the top level of the src
directory hierarchty. Although most end to end tests would reside
there, there are tests than are more narrowly focused that can
still benefit from using the e2e machinery, such as the upcoming
test of voila. We would not want such tests to be outside of the
voila directory.

Conversely, there may be other kinds of tests with a broad scope
than those using the end to end machinery, and they should equally
reside at the top of the src directory hierarchy.

Change-Id: Iaa19cb4f708c0d548a68e476958e5ebc7966e6be