[debugger] Separate out identifier base class.
This is the first step in splitting out a full parsed C++ identifier
with template parameters from the general identifier code in the symbols
directory. This will allow us to make identifiers from DWARF symbols
without doing any transformations, while parsing C++ templates in the
few places they're needed. That capability is in turn required to look
up the definitions of types whose variables are defined in DWARF only as
definitions, which is a high priority usability bug.

The identifier is now a templatized base class on a component type.
Currently this uses the same definition as the previous version of a
component. In the future pass the IdentifierComponent will be just a
string and there will be a C++ one with template stuff in it in the
expr/ directory.

The main change is that Identifier::Component -> IdentifierComponent and
some calls were updated to keep them more generic.

There should be no behavior change.

Change-Id: I501bc98237ae47bc72312ea4f9fb8e48fdef07db