[ledger][cloud_provider] AddObject supports sending references
This adds a `references' argument to the PageCloud.AddObject request.
Clients (Ledger and validation tests) do not use this field yet, and
cloud providers do not use it.
Because the list of references may be large, it is passed as a VMO.
The buffer is optional: an absent buffer is treated as an empty
reference list. This is useful because most pieces are chunks of data
instead of indexes, and do not contain any reference.

TEST=no behavior change

Change-Id: I349c6664217449bddc18c3b804c2ca6efefe9d1d