[voila] add a testing session shell
This patch adds a miminal testing session shell implementation, which
only displays "Hello, world" on the screen. It can be used to verify
that view integration is working correctly across Voila and sessionmgr
(catching issues such as LE-746). In a subsequent CL we'll add an SL4F
test based on it.

LE-746

Test: `fx shell tiles_ctl add fuchsia-pkg://fuchsia.com/voila_tests#meta/session_shell.cmx`
Change-Id: Icf5c72c37da1c334ea55af456d106230920c39c2