[netstack] Apply logging rubric to DNS client.
In the case where there is no network connectivity and clients of the
networking stack are attempting to connect to the network in a loop, we
log the "no route" error the resolver gets over and over again at
the WARNING log level.

This doesn't conform to the logging rubric; in this case, the
information that DNS failed is a symptom of a problem upstream of the
DNS client, and this is only of interest to developers of the DNS
client ("this failure is not due to DNS itself").

Clients of DNS downstream can use the error returned from the DNS client
via the AddrInfoStatus value returned from GetAddrInfo to discriminate
between different failure modes.

We continue to log DNS cache fullness at the warning level, since it
should be investigated but does not interfere with the functionality of
the DNS client.

NET-2320 #done

Change-Id: I460d7380a6a787502f248a8396c4833f5855ba64