[amberctl] Mirror source edits to pkg_resolver
This reverts commit 34034708ba7ec01b7e4fccbcd72c6ca578bd0497, which was
a revert of this original change.

amberctl generated the repo config's repo_url field by appending the
source config's ID to "fuchsia-pkg://", but when the source config
contained characters that are not part of a host name (like "http://"),
the repository manager would reject the repo config, causing amberctl to
bail out.

In addition to the original change, this change sanitizes
source config ID to contain only valid characters and adds test cases to
verify the functionality.

Test: amberctl-tests, add-update-source without -n parameter

Change-Id: I2493a91d8806f30c55fd1996d306471c2ef9b1ff