[zxcrypt] Isolate tests from system device manager
To improve testability, zxcrypt's FdioVolume now requires the injection
of an FD representing the root of the devfs tree.  This is needed to be
able to interact with the multiple device drivers that comprise the
zxcrypt system in an isolated devmgr instance.

Since some of the zxcrypt tests also exercise FVM, it was necessary to
add similar devfs-root-relative API to FVM.  I shied away from
converting all of FVM to using only devfs-root-relative paths myself.
It's possible to do, but winds up needing fairly broad changes across
everything that interacts with filesystems and lots of changes to other
tests that currently rely on using the systemwide /dev.

ZX-3430 #done
FLK-226 #done
FLK-235 #done

Change-Id: I0c18dbf3f8c324cbdaf61e7aeedeef027129f4cf