[mediacodec] Begin stress tests of codecs.
This CL begins a test suite of StreamProcessor
implementations.

Testing a StreamProcessor is hard because there
is a lot to keep track of. This CL features a
StreamRunner that takes a description of
intended behavior in data, so that tests can just
specify a config for the stream run.

StreamRunner is generic over elementary stream.
All that is required of new codecs to fit into
the StreamRunner is implementing the
ElementaryStream trait for some test data stream.

A test spec type is defined that allows testers
to specify the elementary streams to run,
the options with which to run them, and a set of
validators for the behavior of the stream.

This is all needed because the scenarios we need
to run for stress tests involve complex scenarios
with many concurrent streams. To keep that under
control it is easiest to define them
declaratively.

Change-Id: Iab9d27448166881ca37fbbe64fb44c4f2495c967