[astro][display] Increase LCD power down/up delay
ZX-4112

Current delays used for LCD power on and off seem to be too short and
does not allow the LCD to completely power off and therefore
communication with LCD becomes impossible. This CL increases the delay
to match the recommended delay values from vendor

Test: Manually power cycle the units more than 100 times. display
initialization succeeded. Furthermore, called DisplayInit in a loop with
and without the change and confirmed the issue does not occur with the
new delay values.

Change-Id: Ib2d02442329543f762c3cee4d6b5dac2dc2efbfd