[debugger] Add identifiers to symbols.
Adds the ability to get the name of a Symbol object as an Identifier.
This allows the exact DWARF symbol structure to be retrieved.

This is necessary to look up symbol names for declarations which is a
high priority bug affecting the ability to print common types when
testing and debugging.

Not all symbols have Identifier names() as described in
Symbol::GetIdentifier().

Incidentally fixes a bug where typedefs did not have the correct
namespace qualification. The test is updated to reflect this. Other than this
there should be no behavior change.

Change-Id: I50745ea7ecbf4fcf6904f2af4e355b7d597be466