[scripts][fuzzing] Fix pattern matching for substrings
Host.filter's pattern matching of fuzzer packages and names was a bit
too aggressive.  If one name was a substring of another, there was no
way to select it without the scripts complain about multiple matches.
This CL modifies the behavior to check for an exact match before attempt
a fuzzy one (pun intended).

Bug: SEC-291
Test: fx run-host-tests fuzzing_host_test
Change-Id: I300d9b8d915750ed0a7f3db06c6ff6ca425748da