[fake-bti] Adding vmo_create_contiguous fake
I added zx_vmo_create_contiguous to the fake-bti
implementation.  The zx_vmo_create_contiguous function
checks if the bti is a valid fake bti, and that the vmo
is valid.  The vmo created is done using vmo_create.
Currently there is no way to make sure that a bti_pin
called with the flag ZX_BTI_CONTIGUOUS only succeeds if
the vmo is created with zx_vmo_create_contiguous.

Also included: a bug fix to the fake-bti library.
The first handle fake_bti_create gave was == ZX_INVALID_HANDLE.
This fix makes the first handle == 2, not 0.

Also, added unit tests to test the above bug, as well as
the zx_vmo_create_contiguous function.

Test: see zircon/system/dev/lib/fake-bti/test/TESTING.MD

Change-Id: Ib7e6b6cee9bb63b218e94f9702c47dd1b108e87a