[debugger] Fix Android unwinder return addresses.
The code we copied from AOSP for Android addresses seems to be wanting
to specify the frame address as being that of the caller of the
function. But the debugger expects the IP to be the return address.

Fixing this just means removing the IP offset code. Additional comments
were added for ARM and x86 about why we have no offsets.

Wording of an internal log message was improved so as not to be
misleading.

TEST=None. This will be tested once we have a checked-in crash dump
test, but that work is in progress.

Change-Id: I9db0c7e443e4e0c08ce8ea00269bc501393070cb