[identity] Initial inspect implementation for account_manager.
This is open loop at the moment but I've tested manually. We're working
on a mechanism for tests to read out of the inspect VMO, at which point
I'll be able to come back and improve test coverage by verifying that
these metrics are set to the values we expect.

AUTH-209 #comment

Change-Id: I2d9e7c8b8f109f471c6bb4a60bcd97ecc5f3a035