[ledger] Update LedgerMemoryEstimator to find the Ledger used in the running benchmark.
Test: LedgerMemoryUsage.{Simple, LaunchTwoLedgers}

Currently, LedgerMemoryEstimator searches for any Ledger instances running
on the device and fails if more than one are running. After this CL, it only
searches for the Ledger component that was started at the same environment
as the benchmark, meaning that look up will succeed even if another Ledger
instance is running (outside the benchmark).

Change-Id: I104121df889cd4d0573326033ef6646036667a98