[system][scsilib] Add unit test of BlockImplQueue sending READS
Add a basic unit test of BlockImplQueue; issue one read to it and
check that the read returns data properly generated by a fake SCSI
Controller. The added test covers the non-zero-copy (non-mapping)
data path on reads, SCSI CDB encoding for reads, and invocation of
the completion callback. It also serves as slight documentation as to
how to use scsi::Disk.

Slightly refactor the unit test to make it easier/clearer to add tests
for asynchronous behavior and other read commands too.

ZX-2314
ZX-3757

Change-Id: Iabb8e2581c29129a560ddd5fde8934371b8fa53d