[mediaplayer] fix z-fighting and clipping
This CL addresses a z-fighting issue in which the video shapenode
had the same z value as its parent and a sibling node used for clip
geometry. This fix was to make the video shapenode's z value -1
rather than 0.

Also, clipping wasn't working, because the deprecated SetClip no
longer works. This has been replaced with a call to SetClipPlanes.

THERE IS NO AUTOMATED TEST TO PREVENT REGRESSION, as this would
entail capturing the screen post-scenic.

TEST: fx shell present_view \
  fuchsia-pkg://fuchsia.com/mediaplayer_test_util#meta/mediaplayer_test_util.cmx \
  --loop \
  https://ia800903.us.archive.org/4/items/BigBugBunny_AVCHD/bigBuckBunny_avchd_720p.mp4

MTWN-298 #done

Change-Id: I721b86ca0abfa72ea1c87c9aa85850322b951e80