[dart_test] Set sources_required = false
DNO-509 #comment
BLD-330 #comment

Sources on dart_test and related targets are intended to indicate entrypoints
to allow multiple test targets to be defined with potentially different
environment characteristics (e.g. host-side vs. target-side). sources_required
is intended to placate the SDK roll, so setting it to false for tests should
have no adverse effect.

Also tweaked the deps variable to not die on omission. Omitting
//third_party/dart-pkg/pub/test is usually an error, but the gn failure

  ERROR at //build/dart/test.gni:149:5: Undefined identifier.
      deps += [ ":$_snapshot_target_name" ]
      ^---
  See //build/dart/test/BUILD.gn:7:1: whence it was called.

was relatively unhelpful in that regard.

Test: https://fuchsia-review.googlesource.com/283437
Change-Id: I7c1f274ed4f8ec9a203c61ea274c5b0d48b40421