[zircon][input] When parsing kbd HID reports, ignore rollover errors
This was manifesting as a bug where if you button mashed a lot in
Chrome, the text in the text field would occasionally delete itself.
On closer inspection, when button mashing, I discovered the input method
service would get an event with the 'left command down' modifier, and if
the user pressed 'a' next, it would briefly select all the text in the
field, causing it to get deleted on the next keypress.

It turns out Device Class Definition for Human Interface Devices (HID)
Version 1.11, Appendix C mentions a special case for rollover errors! A
keyboard reports rollover errors by setting all array fields to
ErrorRollOver, which is 1. So the HID report buffer for this error
looked like:

    [1, 0, 1, 1, 1, 1, 1, 1]

hid_kbd_parse_report was not aware of this special case, and although it
*would* set the keymask to say the only key pressed was
HID_USAGE_KEY_ERROR_ROLLOVER, it would also say that this key had a
control key modifier pressed, since in any other situation, a '1' in the
first byte indicates the left control modifier key is pressed. Then, if
the user pressed the 'a' key next, it would report to applications as
control+a pressed, followed by 'control' released.

This change updates the ParseKeyboardReport function to maintain the
previous keymask if the keyboard reports HID_USAGE_KEY_ERROR_ROLLOVER.
This seems better than reporting 'no keys pressed', since then, holding
'A' and then mashing a bunch of keys with your other hand could result
in 'A' registering as being pressed many times, any time there was a
rollover error.

In Virtcon, it also updates vs_input_process to not update pressed keys
if the keyboard reports HID_USAGE_KEY_ERROR_ROLLOVER.

Tests have been added to both virtcon and input_reader.

Change-Id: I534550883458bd4c2f0d96a871d8ad2027b7ab48