[kernel][exceptions] fix thread kill/exception race
Previously, if a caller killed a thread then immediately resumed from
exception (e.g. via closing the exception port), there was a chance
that the resume status would take precedence and we would keep
processing the exception on a killed thread.

This CL adds an additional check so that if the thread is both resumed
and killed, the kill will always take precedence.

This is difficult to test in an automated unittest so these new tests
just run for 1 second and it'll show up as flake in the bots if the bug
returns. To test manually, insert a sleep in the kernel between sending
the exception and waiting for the response to lengthen the race window.

ZX-4105

Test: runtests
Test: while runtests -t exception-test; do :; done
Change-Id: I3894f0141e7dbcc84242aee1583a031028212f71