[elflib] Support reading build ID from sections
To handle split debug binaries we should first read the build ID from the
section and then only if the section is missing fallback to the program
header. I've added tests here for the only section case and the only
program header case but the symbolizer has more tricky tests using
both.

Change-Id: I593e2f4f9d8e3f9ab40e431ebc198ba4e8860a84