[netstack] Fix: connect() sometimes doesn't return
Synchronous connect() internally waits for ZXSIO_SIGNAL_OUTGOING sent
from netstack, but if the connection is shut down on the other end
immediately after the connection is established, netstack fails
to send ZXSIO_SIGNAL_OUTGOING. It causes connect() to be stuck forever.

The problem manifested as a test flake of netstack_bsdsocket_c_api_test
(NetStreamTest.Shutdown), which was reported on NET-2315.

NET-2315 #done

Change-Id: I516cd20fc671680de5a3e6c960dfceadb25e35ac