[fdio] Make fdio_pipe_half match fdio_pipe_half2
This CL is part of a soft transition to change the signature for
fdio_pipe_half. All the clients have moved to fdio_pipe_half2, so we can
now change fdio_pipe_half to the new signature.

This CL removes fdio_pipe_half from the list of untested functions in
FDIO, which seems like a better way to remove this function from that
list than to actually write a test for it.

ZX-4188 #comment

Test: No behavior change because all clients call fdio_pipe_half2 now.
Change-Id: I8420b8bca8b7df20efe6e7e6f75bff76fb512130