[fdio] Delete fdio_ns_install
In trying to write a test for fdio_ns_install, I realized that the
function is essentially useless. In order to call the function
successfully, the process needs to have been created without a
namespace, which is something we did in the past but no longer do.

The function has no callers, so this CL removes it rather than writing a
test for it, which removes this function from the list of FDIO functions
that lack a test.

ZX-4188 #comment

Change-Id: Ieb8a7e39300093e0ea723b3ebcccb7be73f8bb5d