[kernel][x86] Move spinlock to C++
Moves the asm spinlock implementation to C++. The compiler produced code
for these is almost identical to hand written asm with the following
exceptions:
 * Loads, stores and arithmetic now take advantage of the implicit
   zero extension of the high 32-bits of registers when using 32-bit
   operands. This saves a few bytes on not needing the REX prefix.
 * arch_spin_lock keeps a zero register around and every loop iteration
   copies this to eax instead of re-zeroing eax. This is more efficient
   and the extra instruction in the uncontended path creates no
   measurable slowdown.

Motivation for migrating the implementation to C++ is to allow for
easier modifications and additions, specifically the percpu counting
number of spinlocks held.

This change produces no discernable performance change in the spinlock
benchmarks in 'k bench'.

Change-Id: Iba2ee4ea1576d04ae52322ca4f120aa1488c9d6b