[gfx] Fix screenshotter utility for non-UMA devices
DX-1517 #comment

This code was assuming that host-visible linear buffers can
always be used as color attachments. That's often not the case
on discrete GPUs and never supported by goldfish vulkan.

This updates the code to use optimal tiling for the image
and a staging buffer to convert the image to a format
that can be read by the CPU.

The result is that tests that rely on screenshotter can run
on AEMU and non-UMA devices in general.

Test: fx shell run fuchsia-pkg://fuchsia.com/scenic_tests#meta/gfx_pixeltests.cmx
Change-Id: I757b4809e68966306813aac51362edccf9066a5b