[trace-reader] Fix non-constant expression in static_assert
A static_cast to a pointer is not allowed in a constant expression.
The argument to fbl::count_of is actually in an evaluated context
and so has to be a proper constant expression, even though the
function itself doesn't use the argument so it's as if it were
unevaluated for most purposes.  Newer compilers can be more strict
about these rules.

Change-Id: Ic032707625d07c529ceca831b21f99b44c909540