[zircon][aml-sd-emmc] Move to Hwreg
In preparation to writing unit tests, this change moves aml-sd-emmc to
use the hwreg library.

Test: Astro wifi. wlan scan;wlan connect;ping www.google.com
More testing coming. To keep changes small, Unit tests will be written in a separate change
and the link will be pasted here. My goal is to get the review
simultaneously, but I will be submitting only after unit tests.

Change-Id: I0e22647cca6dd75342ef5963091a47c71839a984