[amber] Use tuf.Client.Target to fetch target metadata
In the upcoming go-tuf G2 migration, paths in the targets role no longer
start with a leading path separator. This allows amber's MerkleFor
function to be portable across go-tuf G1.5 and G2.

Tests: There is no direct behavior change in amber, this just moves this
functionality into go-tuf (which is in the third_party/go-tuf
repository). This functionality is indirectly tested by the amber daemon
tests.
Bug: PKG-754 #done

Change-Id: Ia527320779234dbd20d5fe080e43171553b8eee2