[debugger] Use registers on the StackFrame.
Previously the registers must be requested asynchonously when needed by
the expression evaluator. Now that the registers are on each stack
frame, they can be returned directly which works for non-topmost stack
frames.

Removes the restriction that the frame registers did not include IP and
SP. This just complicated the code even though the values are
technically duplicated.

Because all general registers are synchronous now, the test for an
asynchronous base pointer was rewritten to refer to memory (always
requires an async call) rather than a register.

Separates out the memory lookup mocking from MockSymbolDataProvider so
the logic can be shared with the MockRemoteAPI to support the test.

Adds DWARF register mappings for x86 xmm0->xmm15. These are not
currently used. This was implemented because I was going to use them for
a test that ended up being written another way. We'll need them
eventually so I left them in with a test.

Change-Id: I25da76d30628095c5355258845596f47a0888621