[kernel][arm64] Replace arm_boot_cpu_lock with atomics
During boot, the secondary CPUs need to wait for the primary to
complete some initialization before they can proceed.  Prior to this
change, it was accomplished using a special spinlock that started in
the locked state (arm_boot_cpu_lock).  Now, we use an atomic integer
to signal that the secondary CPUs may proceed.

Why change it?  Separately, spinlocks are being enhanced to keep track
of the number of spinlocks held (ZX-147).  For this use case, we don't
really need a spinlock to synchronize with the secondary CPUs and the
specialized nature of arm_boot_cpu_lock would require additional logic
to ensure the spinlock count is accurate.

Test: repeatedly booted vim2, qemu/arm64, and qemu/arm64+kvm(x2)
Change-Id: Id1f08213c7c09d61f40a1ec41f52038cbec90f24