[mediasession][rust][fidl] Move mediasession test to into_ methods
These new convenience methods on the generated Request and Event
types allow for more ergonomic testing. This change moves the
mediasession integration test to use these functions.

Additionally, this change removes the tupling of single-value
event and request objects in the into_ methods to prevent having
to unpack single values.

Change-Id: I182a2e57faffa5ee921b91742e85cd298504dba7