[netstack] Panic given invalid input in AddRoutesLocked
...enabling syzkaller to find the routing table state management bug
that's the root cause for NET-2244.

Once we get a call stack or a reduced reproducer from syzkaller, we can
fix NET-2244 and return an error from AddRoutesLocked instead.

This patch found a bug in util.ApplyMask - if mask was zero-length or
unset (which it happened to be in responses from the dhcp server in
netstack_fidl_integration_lib_test#acquire_dhcp) we would silently
return a zero-length address as the result.

This also revealed a bug in our handling of DHCP address acquisition -
the subnet mask option is not guaranteed to be specified by DHCP servers
(RFC 2131, Table 3), and we were unconditionally using it to compute
additional routes for the default route table.

We'll instead default to the class-based subnet mask for the offered
address.

Testing all this required the first tests of our DHCP address
acquisition logic, so I went ahead and wrote tests for all the expected
cases for DHCP acquisition route table manipulation.

NET-2323 #done

Change-Id: I3b7538baa38246c936ef7a92044014ac963f6292