[fs-test-utils] bloblist truncate should ignore ENOSPC
This changes ConfigBlob on BlobList to handle ENOSPC errors by putting
the unconfigured blob back onto the list instead of erroring out.

Not worrying about running out of space is fine for the tests that the
bloblist is currently being used for, where the number of operations is
strictly limited and we are operating on a ramdisk that can take up all
of memory. This change has no functional impact for those. But a new
test coming down the pipeline may operate for an unspecified amount of
time, and on a physical partition with a limited size. In essense, we
are preventing future flakes.

Test=no runtime changes

Change-Id: I9d6c8d9abfe80f7cb5333a52b6fb74e9941b1ea1