[zircon][dev][goldfish] Add goldfish control driver
DX-939 #comment

This implements a goldfish control device driver. The
purpose of this device is to manage HW side resources
associated sysmem allocated VMOs. It provides a FIDL
interface that can be used by the vulkan ICD to create
HW resources (color buffers) for a VMO and query the
existing resource associated with a VMO.

In order to support this VMO to color buffer mapping,
the concept of a 'heap' is introduced in sysmem. Sysmem
participants can require that memory is allocated on
a specific heap. Memory allocated on a specific heap
can be device local and doesn't need to support CPU
access.

The FIDL interface makes it possible to implement
import/export of memory dedicated to images in the
goldfish vulkan ICD.

The query interface is also exposed as a banjo protocol.
This allows child drivers (e.g. goldfish display driver)
to acquire HW resource associated with VMOs.

This change is part of a series of changes that allow
Fuchsia to run in AEMU, which makes it possible to run a
large set of our existing UI unit/integration tests
without Fuchsia hardware.

Test: /boot/test/sys/sysmem-test
Test: /boot/test/sys/golfish-test
Change-Id: Ib3517bcbf4f5b8561f51f6a7cb19805d984f7923