[port] remove incorrect assertion in CancelQueued
ZX-4206 is cause by an incorrect assertion introduced in a previous
change, that assumed that packets removed by CancelQueued cannot be
ephemeral. The reasoning was that CancelQueued is used with a handle
and ephemeral packets cannot contain a handle.
But CancelQueued is also used for packets added by hypervisor traps,
and these packets are ephemeral and not associated to a handle, thus
CancelQueued is called with a null handle, and we panic.
This change removes the assertion.

TEST=added guest_set_trap_with_bell_drop, crashes on patchset #1

Change-Id: Ia6659f4ee4d97780619bead3b8058d61e79e72b8