[mediaplayer] fix panic in mediaplayer_test_util
This CL fixes a panic that occurs in mediaplayer_test_util when
mediaplayer is destroyed by e.g. killall mediaplayer.cmx. The
member variable |video_view_holder_| is cleared when the view
dies and subsequently dereferenced in the Layout method. This
change adds a guard to the top of Layout to prevent it from
running when |video_view_holder_| is null and causes the utility
to quit in this case.

TEST: fx shell present_view \
  fuchsia-pkg://fuchsia.com/mediaplayer_test_util#meta/mediaplayer_test_util.cmx \
  --loop \
  https://ia800903.us.archive.org/4/items/BigBugBunny_AVCHD/bigBuckBunny_avchd_720p.mp4
      fx shell killall mediaplayer.cmx
Change-Id: Ia38a9f0fcdc0ab083f27a56351030fc92ac1014f