[zircon][dev][goldfish] Add goldfish display driver
DX-939 #comment

This adds a goldfish display controller implementation. The
change is a critical step towards AEMU support and testing
Fuchsia UI without access to Fuchsia supported hardware.

The current framebuffer is transferred to host at each refresh
interval so performance is not expected to be great when using
linear buffers. This can be improved by using address space
device and mapping host side memory into physical address space.

Non-linear (Vulkan) buffer handling is more efficient as
uploads to color buffers are not needed.

This change is part of a series of changes that allow
Fuchsia to run in AEMU, which makes it possible to run a
large set of our existing UI unit/integration tests
without Fuchsia hardware.

Test: gfxtest
Test: display-test --simple
Change-Id: Ib68dc1c23e25fcf921102aa6c7765435f18648ba