[modular][testing] Replace story_shell integration test.
This CL does not yet include verification that the story shell gets the
ModuleManifest for the given module. It appears that the TestHarness
fake package loader does not give sessionmgr access to the module
manifest. The existing cmx loader relies on the component to be in a
package.

That will have to come in a follow-up CL.

MF-400 #done

Change-Id: I3ea143722c558996aa87633c56c81418ff78cf9d